Retrieve tibble with function documentation for a specified package

Retrieve tibble with function documentation for a specified package

Usage

forgot(pkg, keyword = NULL, selected = NULL, formatted = T, interactive = F)

Arguments

pkg

string with installed R package name

keyword

optional string to filter the dataframe on for any mention of

selected

optional string vector of columns to search on, null by default

formatted

boolean should it coerce columns to char, true by default

interactive

optional boolean to get a reactable HTML table

Value

tibble with function documentation info

Examples

forgot::forgot("stringr")
#> # A tibble: 37 × 13
#>    function_name title    usage desc  value author examples name  aliases params
#>    <chr>         <chr>    <chr> <chr> <chr> <chr>  <chr>    <chr> <chr>   <chr> 
#>  1 case          Convert… "\ns… "\n\… "\nA… chara… "\ndog … case  "c(\"c… "c(\"…
#>  2 invert_match  Switch … "\ni… "\nI… "\nn… chara… "\nnumb… inve… "NULL"  "loc …
#>  3 modifiers     Control… "\nf… "\nM… "\nA… chara… "\npatt… modi… "c(\"m… "c(\"…
#>  4 %>%           Pipe op… "\nl… "\nP… "NUL… chara… "NULL"   %>%   "NULL"  "NULL"
#>  5 str_c         Join mu… "\ns… "\n\… "\nI… chara… "\nstr_… str_c "NULL"  "c(\"…
#>  6 str_conv      Specify… "\ns… "\nT… "NUL… chara… "\n# Ex… str_… "NULL"  "c(\"…
#>  7 str_count     Count n… "\ns… "\nC… "\nA… chara… "\nfrui… str_… "NULL"  "c(\"…
#>  8 str_detect    Detect … "\ns… "\n\… "\nA… chara… "\nfrui… str_… "NULL"  "c(\"…
#>  9 str_dup       Duplica… "\ns… "\n\… "\nA… chara… "\nfrui… str_… "NULL"  "c(\"…
#> 10 str_equal     Determi… "\ns… "\nT… "\nA… chara… "\n# Th… str_… "NULL"  "c(\"…
#> # ℹ 27 more rows
#> # ℹ 3 more variables: keywords <chr>, seealso <chr>, format <chr>
forgot::forgot("dplyr", "count")
#> # A tibble: 11 × 15
#>    function_name title   usage desc  section value author seealso examples name 
#>    <chr>         <chr>   <chr> <chr> <chr>   <chr> <chr>  <chr>   <chr>    <chr>
#>  1 count         "Count… "\nc… "\n\… "NULL"  "\nA… chara… "NULL"  "\n# co… count
#>  2 distinct_all  "Selec… "\nd… "\n\… "Group… "NUL… chara… "NULL"  "\ndf <… dist…
#>  3 dplyr_by      "Per-o… "NUL… "\nT… "NULL"  "NUL… chara… "NULL"  "NULL"   dply…
#>  4 filter_all    "Filte… "\nf… "\n\… "Group… "NUL… chara… "NULL"  "\n# Wh… filt…
#>  5 if_else       "Vecto… "\ni… "\n\… "NULL"  "\nA… chara… "NULL"  "\nx <-… if_e…
#>  6 n_distinct    "Count… "\nn… "\n\… "NULL"  "\nA… chara… "NULL"  "\nx <-… n_di…
#>  7 percent_rank  "Propo… "\np… "\nT… "NULL"  "\nA… chara… "\nOth… "\nx <-… perc…
#>  8 pick          "Selec… "\np… "\n\… "NULL"  "\nA… chara… "\n\\c… "\ndf <… pick 
#>  9 pull          "Extra… "\np… "\n\… "Metho… "\nA… chara… "NULL"  "\nmtca… pull 
#> 10 se-deprecated "Depre… "\na… "\n\… "NULL"  "NUL… chara… "NULL"  "NULL"   se-d…
#> 11 summarise     "Summa… "\ns… "\n\… "Usefu… "\nA… chara… "\nOth… "\n# A … summ…
#> # ℹ 5 more variables: aliases <chr>, params <chr>, keywords <chr>,
#> #   details <chr>, format <chr>
forgot::forgot(dplyr, "count")
#> # A tibble: 11 × 15
#>    function_name title   usage desc  section value author seealso examples name 
#>    <chr>         <chr>   <chr> <chr> <chr>   <chr> <chr>  <chr>   <chr>    <chr>
#>  1 count         "Count… "\nc… "\n\… "NULL"  "\nA… chara… "NULL"  "\n# co… count
#>  2 distinct_all  "Selec… "\nd… "\n\… "Group… "NUL… chara… "NULL"  "\ndf <… dist…
#>  3 dplyr_by      "Per-o… "NUL… "\nT… "NULL"  "NUL… chara… "NULL"  "NULL"   dply…
#>  4 filter_all    "Filte… "\nf… "\n\… "Group… "NUL… chara… "NULL"  "\n# Wh… filt…
#>  5 if_else       "Vecto… "\ni… "\n\… "NULL"  "\nA… chara… "NULL"  "\nx <-… if_e…
#>  6 n_distinct    "Count… "\nn… "\n\… "NULL"  "\nA… chara… "NULL"  "\nx <-… n_di…
#>  7 percent_rank  "Propo… "\np… "\nT… "NULL"  "\nA… chara… "\nOth… "\nx <-… perc…
#>  8 pick          "Selec… "\np… "\n\… "NULL"  "\nA… chara… "\n\\c… "\ndf <… pick 
#>  9 pull          "Extra… "\np… "\n\… "Metho… "\nA… chara… "NULL"  "\nmtca… pull 
#> 10 se-deprecated "Depre… "\na… "\n\… "NULL"  "NUL… chara… "NULL"  "NULL"   se-d…
#> 11 summarise     "Summa… "\ns… "\n\… "Usefu… "\nA… chara… "\nOth… "\n# A … summ…
#> # ℹ 5 more variables: aliases <chr>, params <chr>, keywords <chr>,
#> #   details <chr>, format <chr>