Print usage for an R function of interest in console

Print usage for an R function of interest in console

Usage

forgot_usg(pkg, function_name, write = FALSE)

Arguments

pkg

string with installed R package name

function_name

function name in R package

write

logical write to RStudio rmd file, FALSE by default

Examples

forgot_usg("dplyr", "count")
#> 
#> across(.cols, .fns, ..., .names = NULL, .unpack = FALSE)
#> 
#> if_any(.cols, .fns, ..., .names = NULL)
#> 
#> if_all(.cols, .fns, ..., .names = NULL)
#>  
#> add_rownames(df, var = "rowname")
#>  
#> all_equal(
#>   target,
#>   current,
#>   ignore_col_order = TRUE,
#>   ignore_row_order = TRUE,
#>   convert = FALSE,
#>   ...
#> )
#>  
#> all_vars(expr)
#> 
#> any_vars(expr)
#>  NULL 
#> arrange(.data, ..., .by_group = FALSE)
#> 
#> \method{arrangedata.frame}(.data, ..., .by_group = FALSE, .locale = NULL)
#>  
#> arrange_all(.tbl, .funs = list(), ..., .by_group = FALSE, .locale = NULL)
#> 
#> arrange_at(.tbl, .vars, .funs = list(), ..., .by_group = FALSE, .locale = NULL)
#> 
#> arrange_if(
#>   .tbl,
#>   .predicate,
#>   .funs = list(),
#>   ...,
#>   .by_group = FALSE,
#>   .locale = NULL
#> )
#>  
#> auto_copy(x, y, copy = FALSE, ...)
#>  
#> db_desc(x)
#> 
#> sql_translate_env(con)
#> 
#> db_list_tables(con)
#> 
#> db_has_table(con, table)
#> 
#> db_data_type(con, fields)
#> 
#> db_save_query(con, sql, name, temporary = TRUE, ...)
#> 
#> db_begin(con, ...)
#> 
#> db_commit(con, ...)
#> 
#> db_rollback(con, ...)
#> 
#> db_write_table(con, table, types, values, temporary = FALSE, ...)
#> 
#> db_create_table(con, table, types, temporary = FALSE, ...)
#> 
#> db_insert_into(con, table, values, ...)
#> 
#> db_create_indexes(con, table, indexes = NULL, unique = FALSE, ...)
#> 
#> db_create_index(con, table, columns, name = NULL, unique = FALSE, ...)
#> 
#> db_drop_table(con, table, force = FALSE, ...)
#> 
#> db_analyze(con, table, ...)
#> 
#> db_explain(con, sql, ...)
#> 
#> db_query_fields(con, sql, ...)
#> 
#> db_query_rows(con, sql, ...)
#> 
#> sql_select(
#>   con,
#>   select,
#>   from,
#>   where = NULL,
#>   group_by = NULL,
#>   having = NULL,
#>   order_by = NULL,
#>   limit = NULL,
#>   distinct = FALSE,
#>   ...
#> )
#> 
#> sql_subquery(con, from, name = random_table_name(), ...)
#> 
#> sql_join(con, x, y, vars, type = "inner", by = NULL, ...)
#> 
#> sql_semi_join(con, x, y, anti = FALSE, by = NULL, ...)
#> 
#> sql_set_op(con, x, y, method)
#> 
#> sql_escape_string(con, x)
#> 
#> sql_escape_ident(con, x)
#>  
#> band_members
#> 
#> band_instruments
#> 
#> band_instruments2
#>  
#> between(x, left, right)
#>  
#> bind_cols(
#>   ...,
#>   .name_repair = c("unique", "universal", "check_unique", "minimal")
#> )
#>  
#> bind_rows(..., .id = NULL)
#>  
#> c_across(cols)
#>  
#> case_match(.x, ..., .default = NULL, .ptype = NULL)
#>  
#> case_when(..., .default = NULL, .ptype = NULL, .size = NULL)
#>  
#> check_dbplyr()
#> 
#> wrap_dbplyr_obj(obj_name)
#>  
#> coalesce(..., .ptype = NULL, .size = NULL)
#>  
#> combine(...)
#>  
#> common_by(by = NULL, x, y)
#>  
#> compute(x, ...)
#> 
#> collect(x, ...)
#> 
#> collapse(x, ...)
#>  
#> consecutive_id(...)
#>  
#> n()
#> 
#> cur_group()
#> 
#> cur_group_id()
#> 
#> cur_group_rows()
#> 
#> cur_column()
#>  
#> copy_to(dest, df, name = deparse(substitute(df)), overwrite = FALSE, ...)
#>  
#> count(x, ..., wt = NULL, sort = FALSE, name = NULL)
#> 
#> \method{countdata.frame}(
#>   x,
#>   ...,
#>   wt = NULL,
#>   sort = FALSE,
#>   name = NULL,
#>   .drop = group_by_drop_default(x)
#> )
#> 
#> tally(x, wt = NULL, sort = FALSE, name = NULL)
#> 
#> add_count(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = deprecated())
#> 
#> add_tally(x, wt = NULL, sort = FALSE, name = NULL)
#>  
#> cross_join(x, y, ..., copy = FALSE, suffix = c(".x", ".y"))
#>  
#> cumall(x)
#> 
#> cumany(x)
#> 
#> cummean(x)
#>  
#> # Deprecated in 0.5.0 -------------------------------------
#> 
#> id(.variables, drop = FALSE)
#> 
#> # Deprecated in 0.7.0 -------------------------------------
#> 
#> failwith(default = NULL, f, quiet = FALSE)
#> 
#> # Deprecated in 0.8.* -------------------------------------
#> 
#> select_vars(vars = chr(), ..., include = chr(), exclude = chr())
#> 
#> rename_vars(vars = chr(), ..., strict = TRUE)
#> 
#> select_var(vars, var = -1)
#> 
#> current_vars(...)
#> 
#> # Deprecated in 1.0.0 -------------------------------------
#> 
#> bench_tbls(tbls, op, ..., times = 10)
#> 
#> compare_tbls(tbls, op, ref = NULL, compare = equal_data_frame, ...)
#> 
#> compare_tbls2(tbls_x, tbls_y, op, ref = NULL, compare = equal_data_frame, ...)
#> 
#> eval_tbls(tbls, op)
#> 
#> eval_tbls2(tbls_x, tbls_y, op)
#> 
#> location(df)
#> 
#> changes(x, y)
#>  
#> cur_data()
#> 
#> cur_data_all()
#>  
#> desc(x)
#>  
#> dim_desc(x)
#>  
#> distinct(.data, ..., .keep_all = FALSE)
#>  
#> distinct_all(.tbl, .funs = list(), ..., .keep_all = FALSE)
#> 
#> distinct_at(.tbl, .vars, .funs = list(), ..., .keep_all = FALSE)
#> 
#> distinct_if(.tbl, .predicate, .funs = list(), ..., .keep_all = FALSE)
#>  
#> do(.data, ...)
#>  NULL NULL NULL NULL 
#> dplyr_row_slice(data, i, ...)
#> 
#> dplyr_col_modify(data, cols)
#> 
#> dplyr_reconstruct(data, template)
#>  NULL 
#> explain(x, ...)
#> 
#> show_query(x, ...)
#>  
#> semi_join(x, y, by = NULL, copy = FALSE, ...)
#> 
#> \method{semi_joindata.frame}(x, y, by = NULL, copy = FALSE, ..., na_matches = c("na", "never"))
#> 
#> anti_join(x, y, by = NULL, copy = FALSE, ...)
#> 
#> \method{anti_joindata.frame}(x, y, by = NULL, copy = FALSE, ..., na_matches = c("na", "never"))
#>  
#> filter(.data, ..., .by = NULL, .preserve = FALSE)
#>  
#> filter_all(.tbl, .vars_predicate, .preserve = FALSE)
#> 
#> filter_if(.tbl, .predicate, .vars_predicate, .preserve = FALSE)
#> 
#> filter_at(.tbl, .vars, .vars_predicate, .preserve = FALSE)
#>  
#> funs(..., .args = list())
#>  NULL 
#> group_by(.data, ..., .add = FALSE, .drop = group_by_drop_default(.data))
#> 
#> ungroup(x, ...)
#>  
#> group_by_all(
#>   .tbl,
#>   .funs = list(),
#>   ...,
#>   .add = FALSE,
#>   .drop = group_by_drop_default(.tbl)
#> )
#> 
#> group_by_at(
#>   .tbl,
#>   .vars,
#>   .funs = list(),
#>   ...,
#>   .add = FALSE,
#>   .drop = group_by_drop_default(.tbl)
#> )
#> 
#> group_by_if(
#>   .tbl,
#>   .predicate,
#>   .funs = list(),
#>   ...,
#>   .add = FALSE,
#>   .drop = group_by_drop_default(.tbl)
#> )
#>  
#> group_by_drop_default(.tbl)
#>  
#> distinct_prepare(
#>   .data,
#>   vars,
#>   group_vars = character(),
#>   .keep_all = FALSE,
#>   caller_env = caller_env(2),
#>   error_call = caller_env()
#> )
#> 
#> group_by_prepare(
#>   .data,
#>   ...,
#>   .add = FALSE,
#>   .dots = deprecated(),
#>   add = deprecated(),
#>   error_call = caller_env()
#> )
#>  
#> group_cols(vars = NULL, data = NULL)
#>  
#> group_data(.data)
#> 
#> group_keys(.tbl, ...)
#> 
#> group_rows(.data)
#> 
#> group_indices(.data, ...)
#> 
#> group_vars(x)
#> 
#> groups(x)
#> 
#> group_size(x)
#> 
#> n_groups(x)
#>  
#> group_map(.data, .f, ..., .keep = FALSE)
#> 
#> group_modify(.data, .f, ..., .keep = FALSE)
#> 
#> group_walk(.data, .f, ..., .keep = FALSE)
#>  
#> group_nest(.tbl, ..., .key = "data", keep = FALSE)
#>  
#> group_split(.tbl, ..., .keep = TRUE)
#>  
#> group_trim(.tbl, .drop = group_by_drop_default(.tbl))
#>  
#> grouped_df(data, vars, drop = group_by_drop_default(data))
#> 
#> is.grouped_df(x)
#> 
#> is_grouped_df(x)
#>  
#> ident(...)
#>  
#> if_else(condition, true, false, missing = NULL, ..., ptype = NULL, size = NULL)
#>  
#> join_by(...)
#>  
#> last_dplyr_warnings(n = 5)
#>  
#> lag(x, n = 1L, default = NULL, order_by = NULL, ...)
#> 
#> lead(x, n = 1L, default = NULL, order_by = NULL, ...)
#>  
#> make_tbl(subclass, ...)
#>  
#> inner_join(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL
#> )
#> 
#> \method{inner_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL,
#>   na_matches = c("na", "never"),
#>   multiple = "all",
#>   unmatched = "drop",
#>   relationship = NULL
#> )
#> 
#> left_join(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL
#> )
#> 
#> \method{left_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL,
#>   na_matches = c("na", "never"),
#>   multiple = "all",
#>   unmatched = "drop",
#>   relationship = NULL
#> )
#> 
#> right_join(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL
#> )
#> 
#> \method{right_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL,
#>   na_matches = c("na", "never"),
#>   multiple = "all",
#>   unmatched = "drop",
#>   relationship = NULL
#> )
#> 
#> full_join(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL
#> )
#> 
#> \method{full_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL,
#>   na_matches = c("na", "never"),
#>   multiple = "all",
#>   relationship = NULL
#> )
#>  
#> mutate(.data, ...)
#> 
#> \method{mutatedata.frame}(
#>   .data,
#>   ...,
#>   .by = NULL,
#>   .keep = c("all", "used", "unused", "none"),
#>   .before = NULL,
#>   .after = NULL
#> )
#>  
#> mutate_all(.tbl, .funs, ...)
#> 
#> mutate_if(.tbl, .predicate, .funs, ...)
#> 
#> mutate_at(.tbl, .vars, .funs, ..., .cols = NULL)
#> 
#> transmute_all(.tbl, .funs, ...)
#> 
#> transmute_if(.tbl, .predicate, .funs, ...)
#> 
#> transmute_at(.tbl, .vars, .funs, ..., .cols = NULL)
#>  
#> n_distinct(..., na.rm = FALSE)
#>  
#> na_if(x, y)
#>  
#> near(x, y, tol = .Machine$double.eps^0.5)
#>  
#> nest_by(.data, ..., .key = "data", .keep = FALSE)
#>  
#> nest_join(x, y, by = NULL, copy = FALSE, keep = NULL, name = NULL, ...)
#> 
#> \method{nest_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   keep = NULL,
#>   name = NULL,
#>   ...,
#>   na_matches = c("na", "never"),
#>   unmatched = "drop"
#> )
#>  
#> new_grouped_df(x, groups, ..., class = character())
#> 
#> validate_grouped_df(x, check_bounds = FALSE)
#> 
#> new_rowwise_df(data, group_data = NULL, ..., class = character())
#> 
#> validate_rowwise_df(x)
#>  
#> nth(x, n, order_by = NULL, default = NULL, na_rm = FALSE)
#> 
#> first(x, order_by = NULL, default = NULL, na_rm = FALSE)
#> 
#> last(x, order_by = NULL, default = NULL, na_rm = FALSE)
#>  
#> ntile(x = row_number(), n)
#>  
#> order_by(order_by, call)
#>  
#> percent_rank(x)
#> 
#> cume_dist(x)
#>  
#> pick(...)
#>  
#> progress_estimated(n, min_time = 0)
#>  
#> pull(.data, var = -1, name = NULL, ...)
#>  
#> recode(.x, ..., .default = NULL, .missing = NULL)
#> 
#> recode_factor(.x, ..., .default = NULL, .missing = NULL, .ordered = FALSE)
#>  NULL 
#> reframe(.data, ..., .by = NULL)
#>  
#> relocate(.data, ..., .before = NULL, .after = NULL)
#>  
#> rename(.data, ...)
#> 
#> rename_with(.data, .fn, .cols = everything(), ...)
#>  
#> row_number(x)
#> 
#> min_rank(x)
#> 
#> dense_rank(x)
#>  
#> rows_insert(
#>   x,
#>   y,
#>   by = NULL,
#>   ...,
#>   conflict = c("error", "ignore"),
#>   copy = FALSE,
#>   in_place = FALSE
#> )
#> 
#> rows_append(x, y, ..., copy = FALSE, in_place = FALSE)
#> 
#> rows_update(
#>   x,
#>   y,
#>   by = NULL,
#>   ...,
#>   unmatched = c("error", "ignore"),
#>   copy = FALSE,
#>   in_place = FALSE
#> )
#> 
#> rows_patch(
#>   x,
#>   y,
#>   by = NULL,
#>   ...,
#>   unmatched = c("error", "ignore"),
#>   copy = FALSE,
#>   in_place = FALSE
#> )
#> 
#> rows_upsert(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)
#> 
#> rows_delete(
#>   x,
#>   y,
#>   by = NULL,
#>   ...,
#>   unmatched = c("error", "ignore"),
#>   copy = FALSE,
#>   in_place = FALSE
#> )
#>  
#> rowwise(data, ...)
#>  
#> same_src(x, y)
#>  
#> sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)
#> 
#> sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)
#>  NULL 
#> add_count_(x, vars, wt = NULL, sort = FALSE)
#> 
#> add_tally_(x, wt, sort = FALSE)
#> 
#> arrange_(.data, ..., .dots = list())
#> 
#> count_(x, vars, wt = NULL, sort = FALSE, .drop = group_by_drop_default(x))
#> 
#> distinct_(.data, ..., .dots, .keep_all = FALSE)
#> 
#> do_(.data, ..., .dots = list())
#> 
#> filter_(.data, ..., .dots = list())
#> 
#> funs_(dots, args = list(), env = base_env())
#> 
#> group_by_(.data, ..., .dots = list(), add = FALSE)
#> 
#> group_indices_(.data, ..., .dots = list())
#> 
#> mutate_(.data, ..., .dots = list())
#> 
#> tally_(x, wt, sort = FALSE)
#> 
#> transmute_(.data, ..., .dots = list())
#> 
#> rename_(.data, ..., .dots = list())
#> 
#> rename_vars_(vars, args)
#> 
#> select_(.data, ..., .dots = list())
#> 
#> select_vars_(vars, args, include = chr(), exclude = chr())
#> 
#> slice_(.data, ..., .dots = list())
#> 
#> summarise_(.data, ..., .dots = list())
#> 
#> summarize_(.data, ..., .dots = list())
#>  
#> select(.data, ...)
#>  
#> select_all(.tbl, .funs = list(), ...)
#> 
#> rename_all(.tbl, .funs = list(), ...)
#> 
#> select_if(.tbl, .predicate, .funs = list(), ...)
#> 
#> rename_if(.tbl, .predicate, .funs = list(), ...)
#> 
#> select_at(.tbl, .vars, .funs = list(), ...)
#> 
#> rename_at(.tbl, .vars, .funs = list(), ...)
#>  
#> intersect(x, y, ...)
#> 
#> union(x, y, ...)
#> 
#> union_all(x, y, ...)
#> 
#> setdiff(x, y, ...)
#> 
#> setequal(x, y, ...)
#> 
#> symdiff(x, y, ...)
#>  
#> slice(.data, ..., .by = NULL, .preserve = FALSE)
#> 
#> slice_head(.data, ..., n, prop, by = NULL)
#> 
#> slice_tail(.data, ..., n, prop, by = NULL)
#> 
#> slice_min(
#>   .data,
#>   order_by,
#>   ...,
#>   n,
#>   prop,
#>   by = NULL,
#>   with_ties = TRUE,
#>   na_rm = FALSE
#> )
#> 
#> slice_max(
#>   .data,
#>   order_by,
#>   ...,
#>   n,
#>   prop,
#>   by = NULL,
#>   with_ties = TRUE,
#>   na_rm = FALSE
#> )
#> 
#> slice_sample(.data, ..., n, prop, by = NULL, weight_by = NULL, replace = FALSE)
#>  
#> sql(...)
#>  
#> src(subclass, ...)
#> 
#> is.src(x)
#>  
#> src_mysql(
#>   dbname,
#>   host = NULL,
#>   port = 0L,
#>   username = "root",
#>   password = "",
#>   ...
#> )
#> 
#> src_postgres(
#>   dbname = NULL,
#>   host = NULL,
#>   port = NULL,
#>   user = NULL,
#>   password = NULL,
#>   ...
#> )
#> 
#> src_sqlite(path, create = FALSE)
#>  
#> src_local(tbl, pkg = NULL, env = NULL)
#> 
#> src_df(pkg = NULL, env = NULL)
#>  
#> src_tbls(x, ...)
#>  
#> starwars
#>  
#> storms
#>  
#> summarise(.data, ..., .by = NULL, .groups = NULL)
#> 
#> summarize(.data, ..., .by = NULL, .groups = NULL)
#>  
#> summarise_all(.tbl, .funs, ...)
#> 
#> summarise_if(.tbl, .predicate, .funs, ...)
#> 
#> summarise_at(.tbl, .vars, .funs, ..., .cols = NULL)
#> 
#> summarize_all(.tbl, .funs, ...)
#> 
#> summarize_if(.tbl, .predicate, .funs, ...)
#> 
#> summarize_at(.tbl, .vars, .funs, ..., .cols = NULL)
#>  
#> summarise_each(tbl, funs, ...)
#> 
#> summarise_each_(tbl, funs, vars)
#> 
#> mutate_each(tbl, funs, ...)
#> 
#> mutate_each_(tbl, funs, vars)
#> 
#> summarize_each(tbl, funs, ...)
#> 
#> summarize_each_(tbl, funs, vars)
#>  
#> tbl(src, ...)
#> 
#> is.tbl(x)
#>  
#> tbl_df(data)
#> 
#> as.tbl(x, ...)
#>  
#> tbl_ptype(.data)
#>  
#> tbl_vars(x)
#> 
#> tbl_nongroup_vars(x)
#>  NULL 
#> top_n(x, n, wt)
#> 
#> top_frac(x, n, wt)
#>  
#> transmute(.data, ...)
#>  
#> vars(...)
#>  
#> with_groups(.data, .groups, .f, ...)
#>  
#> with_order(order_by, fun, x, ...)
forgot_usg(dplyr, count)
#> 
#> across(.cols, .fns, ..., .names = NULL, .unpack = FALSE)
#> 
#> if_any(.cols, .fns, ..., .names = NULL)
#> 
#> if_all(.cols, .fns, ..., .names = NULL)
#>  
#> add_rownames(df, var = "rowname")
#>  
#> all_equal(
#>   target,
#>   current,
#>   ignore_col_order = TRUE,
#>   ignore_row_order = TRUE,
#>   convert = FALSE,
#>   ...
#> )
#>  
#> all_vars(expr)
#> 
#> any_vars(expr)
#>  NULL 
#> arrange(.data, ..., .by_group = FALSE)
#> 
#> \method{arrangedata.frame}(.data, ..., .by_group = FALSE, .locale = NULL)
#>  
#> arrange_all(.tbl, .funs = list(), ..., .by_group = FALSE, .locale = NULL)
#> 
#> arrange_at(.tbl, .vars, .funs = list(), ..., .by_group = FALSE, .locale = NULL)
#> 
#> arrange_if(
#>   .tbl,
#>   .predicate,
#>   .funs = list(),
#>   ...,
#>   .by_group = FALSE,
#>   .locale = NULL
#> )
#>  
#> auto_copy(x, y, copy = FALSE, ...)
#>  
#> db_desc(x)
#> 
#> sql_translate_env(con)
#> 
#> db_list_tables(con)
#> 
#> db_has_table(con, table)
#> 
#> db_data_type(con, fields)
#> 
#> db_save_query(con, sql, name, temporary = TRUE, ...)
#> 
#> db_begin(con, ...)
#> 
#> db_commit(con, ...)
#> 
#> db_rollback(con, ...)
#> 
#> db_write_table(con, table, types, values, temporary = FALSE, ...)
#> 
#> db_create_table(con, table, types, temporary = FALSE, ...)
#> 
#> db_insert_into(con, table, values, ...)
#> 
#> db_create_indexes(con, table, indexes = NULL, unique = FALSE, ...)
#> 
#> db_create_index(con, table, columns, name = NULL, unique = FALSE, ...)
#> 
#> db_drop_table(con, table, force = FALSE, ...)
#> 
#> db_analyze(con, table, ...)
#> 
#> db_explain(con, sql, ...)
#> 
#> db_query_fields(con, sql, ...)
#> 
#> db_query_rows(con, sql, ...)
#> 
#> sql_select(
#>   con,
#>   select,
#>   from,
#>   where = NULL,
#>   group_by = NULL,
#>   having = NULL,
#>   order_by = NULL,
#>   limit = NULL,
#>   distinct = FALSE,
#>   ...
#> )
#> 
#> sql_subquery(con, from, name = random_table_name(), ...)
#> 
#> sql_join(con, x, y, vars, type = "inner", by = NULL, ...)
#> 
#> sql_semi_join(con, x, y, anti = FALSE, by = NULL, ...)
#> 
#> sql_set_op(con, x, y, method)
#> 
#> sql_escape_string(con, x)
#> 
#> sql_escape_ident(con, x)
#>  
#> band_members
#> 
#> band_instruments
#> 
#> band_instruments2
#>  
#> between(x, left, right)
#>  
#> bind_cols(
#>   ...,
#>   .name_repair = c("unique", "universal", "check_unique", "minimal")
#> )
#>  
#> bind_rows(..., .id = NULL)
#>  
#> c_across(cols)
#>  
#> case_match(.x, ..., .default = NULL, .ptype = NULL)
#>  
#> case_when(..., .default = NULL, .ptype = NULL, .size = NULL)
#>  
#> check_dbplyr()
#> 
#> wrap_dbplyr_obj(obj_name)
#>  
#> coalesce(..., .ptype = NULL, .size = NULL)
#>  
#> combine(...)
#>  
#> common_by(by = NULL, x, y)
#>  
#> compute(x, ...)
#> 
#> collect(x, ...)
#> 
#> collapse(x, ...)
#>  
#> consecutive_id(...)
#>  
#> n()
#> 
#> cur_group()
#> 
#> cur_group_id()
#> 
#> cur_group_rows()
#> 
#> cur_column()
#>  
#> copy_to(dest, df, name = deparse(substitute(df)), overwrite = FALSE, ...)
#>  
#> count(x, ..., wt = NULL, sort = FALSE, name = NULL)
#> 
#> \method{countdata.frame}(
#>   x,
#>   ...,
#>   wt = NULL,
#>   sort = FALSE,
#>   name = NULL,
#>   .drop = group_by_drop_default(x)
#> )
#> 
#> tally(x, wt = NULL, sort = FALSE, name = NULL)
#> 
#> add_count(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = deprecated())
#> 
#> add_tally(x, wt = NULL, sort = FALSE, name = NULL)
#>  
#> cross_join(x, y, ..., copy = FALSE, suffix = c(".x", ".y"))
#>  
#> cumall(x)
#> 
#> cumany(x)
#> 
#> cummean(x)
#>  
#> # Deprecated in 0.5.0 -------------------------------------
#> 
#> id(.variables, drop = FALSE)
#> 
#> # Deprecated in 0.7.0 -------------------------------------
#> 
#> failwith(default = NULL, f, quiet = FALSE)
#> 
#> # Deprecated in 0.8.* -------------------------------------
#> 
#> select_vars(vars = chr(), ..., include = chr(), exclude = chr())
#> 
#> rename_vars(vars = chr(), ..., strict = TRUE)
#> 
#> select_var(vars, var = -1)
#> 
#> current_vars(...)
#> 
#> # Deprecated in 1.0.0 -------------------------------------
#> 
#> bench_tbls(tbls, op, ..., times = 10)
#> 
#> compare_tbls(tbls, op, ref = NULL, compare = equal_data_frame, ...)
#> 
#> compare_tbls2(tbls_x, tbls_y, op, ref = NULL, compare = equal_data_frame, ...)
#> 
#> eval_tbls(tbls, op)
#> 
#> eval_tbls2(tbls_x, tbls_y, op)
#> 
#> location(df)
#> 
#> changes(x, y)
#>  
#> cur_data()
#> 
#> cur_data_all()
#>  
#> desc(x)
#>  
#> dim_desc(x)
#>  
#> distinct(.data, ..., .keep_all = FALSE)
#>  
#> distinct_all(.tbl, .funs = list(), ..., .keep_all = FALSE)
#> 
#> distinct_at(.tbl, .vars, .funs = list(), ..., .keep_all = FALSE)
#> 
#> distinct_if(.tbl, .predicate, .funs = list(), ..., .keep_all = FALSE)
#>  
#> do(.data, ...)
#>  NULL NULL NULL NULL 
#> dplyr_row_slice(data, i, ...)
#> 
#> dplyr_col_modify(data, cols)
#> 
#> dplyr_reconstruct(data, template)
#>  NULL 
#> explain(x, ...)
#> 
#> show_query(x, ...)
#>  
#> semi_join(x, y, by = NULL, copy = FALSE, ...)
#> 
#> \method{semi_joindata.frame}(x, y, by = NULL, copy = FALSE, ..., na_matches = c("na", "never"))
#> 
#> anti_join(x, y, by = NULL, copy = FALSE, ...)
#> 
#> \method{anti_joindata.frame}(x, y, by = NULL, copy = FALSE, ..., na_matches = c("na", "never"))
#>  
#> filter(.data, ..., .by = NULL, .preserve = FALSE)
#>  
#> filter_all(.tbl, .vars_predicate, .preserve = FALSE)
#> 
#> filter_if(.tbl, .predicate, .vars_predicate, .preserve = FALSE)
#> 
#> filter_at(.tbl, .vars, .vars_predicate, .preserve = FALSE)
#>  
#> funs(..., .args = list())
#>  NULL 
#> group_by(.data, ..., .add = FALSE, .drop = group_by_drop_default(.data))
#> 
#> ungroup(x, ...)
#>  
#> group_by_all(
#>   .tbl,
#>   .funs = list(),
#>   ...,
#>   .add = FALSE,
#>   .drop = group_by_drop_default(.tbl)
#> )
#> 
#> group_by_at(
#>   .tbl,
#>   .vars,
#>   .funs = list(),
#>   ...,
#>   .add = FALSE,
#>   .drop = group_by_drop_default(.tbl)
#> )
#> 
#> group_by_if(
#>   .tbl,
#>   .predicate,
#>   .funs = list(),
#>   ...,
#>   .add = FALSE,
#>   .drop = group_by_drop_default(.tbl)
#> )
#>  
#> group_by_drop_default(.tbl)
#>  
#> distinct_prepare(
#>   .data,
#>   vars,
#>   group_vars = character(),
#>   .keep_all = FALSE,
#>   caller_env = caller_env(2),
#>   error_call = caller_env()
#> )
#> 
#> group_by_prepare(
#>   .data,
#>   ...,
#>   .add = FALSE,
#>   .dots = deprecated(),
#>   add = deprecated(),
#>   error_call = caller_env()
#> )
#>  
#> group_cols(vars = NULL, data = NULL)
#>  
#> group_data(.data)
#> 
#> group_keys(.tbl, ...)
#> 
#> group_rows(.data)
#> 
#> group_indices(.data, ...)
#> 
#> group_vars(x)
#> 
#> groups(x)
#> 
#> group_size(x)
#> 
#> n_groups(x)
#>  
#> group_map(.data, .f, ..., .keep = FALSE)
#> 
#> group_modify(.data, .f, ..., .keep = FALSE)
#> 
#> group_walk(.data, .f, ..., .keep = FALSE)
#>  
#> group_nest(.tbl, ..., .key = "data", keep = FALSE)
#>  
#> group_split(.tbl, ..., .keep = TRUE)
#>  
#> group_trim(.tbl, .drop = group_by_drop_default(.tbl))
#>  
#> grouped_df(data, vars, drop = group_by_drop_default(data))
#> 
#> is.grouped_df(x)
#> 
#> is_grouped_df(x)
#>  
#> ident(...)
#>  
#> if_else(condition, true, false, missing = NULL, ..., ptype = NULL, size = NULL)
#>  
#> join_by(...)
#>  
#> last_dplyr_warnings(n = 5)
#>  
#> lag(x, n = 1L, default = NULL, order_by = NULL, ...)
#> 
#> lead(x, n = 1L, default = NULL, order_by = NULL, ...)
#>  
#> make_tbl(subclass, ...)
#>  
#> inner_join(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL
#> )
#> 
#> \method{inner_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL,
#>   na_matches = c("na", "never"),
#>   multiple = "all",
#>   unmatched = "drop",
#>   relationship = NULL
#> )
#> 
#> left_join(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL
#> )
#> 
#> \method{left_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL,
#>   na_matches = c("na", "never"),
#>   multiple = "all",
#>   unmatched = "drop",
#>   relationship = NULL
#> )
#> 
#> right_join(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL
#> )
#> 
#> \method{right_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL,
#>   na_matches = c("na", "never"),
#>   multiple = "all",
#>   unmatched = "drop",
#>   relationship = NULL
#> )
#> 
#> full_join(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL
#> )
#> 
#> \method{full_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   suffix = c(".x", ".y"),
#>   ...,
#>   keep = NULL,
#>   na_matches = c("na", "never"),
#>   multiple = "all",
#>   relationship = NULL
#> )
#>  
#> mutate(.data, ...)
#> 
#> \method{mutatedata.frame}(
#>   .data,
#>   ...,
#>   .by = NULL,
#>   .keep = c("all", "used", "unused", "none"),
#>   .before = NULL,
#>   .after = NULL
#> )
#>  
#> mutate_all(.tbl, .funs, ...)
#> 
#> mutate_if(.tbl, .predicate, .funs, ...)
#> 
#> mutate_at(.tbl, .vars, .funs, ..., .cols = NULL)
#> 
#> transmute_all(.tbl, .funs, ...)
#> 
#> transmute_if(.tbl, .predicate, .funs, ...)
#> 
#> transmute_at(.tbl, .vars, .funs, ..., .cols = NULL)
#>  
#> n_distinct(..., na.rm = FALSE)
#>  
#> na_if(x, y)
#>  
#> near(x, y, tol = .Machine$double.eps^0.5)
#>  
#> nest_by(.data, ..., .key = "data", .keep = FALSE)
#>  
#> nest_join(x, y, by = NULL, copy = FALSE, keep = NULL, name = NULL, ...)
#> 
#> \method{nest_joindata.frame}(
#>   x,
#>   y,
#>   by = NULL,
#>   copy = FALSE,
#>   keep = NULL,
#>   name = NULL,
#>   ...,
#>   na_matches = c("na", "never"),
#>   unmatched = "drop"
#> )
#>  
#> new_grouped_df(x, groups, ..., class = character())
#> 
#> validate_grouped_df(x, check_bounds = FALSE)
#> 
#> new_rowwise_df(data, group_data = NULL, ..., class = character())
#> 
#> validate_rowwise_df(x)
#>  
#> nth(x, n, order_by = NULL, default = NULL, na_rm = FALSE)
#> 
#> first(x, order_by = NULL, default = NULL, na_rm = FALSE)
#> 
#> last(x, order_by = NULL, default = NULL, na_rm = FALSE)
#>  
#> ntile(x = row_number(), n)
#>  
#> order_by(order_by, call)
#>  
#> percent_rank(x)
#> 
#> cume_dist(x)
#>  
#> pick(...)
#>  
#> progress_estimated(n, min_time = 0)
#>  
#> pull(.data, var = -1, name = NULL, ...)
#>  
#> recode(.x, ..., .default = NULL, .missing = NULL)
#> 
#> recode_factor(.x, ..., .default = NULL, .missing = NULL, .ordered = FALSE)
#>  NULL 
#> reframe(.data, ..., .by = NULL)
#>  
#> relocate(.data, ..., .before = NULL, .after = NULL)
#>  
#> rename(.data, ...)
#> 
#> rename_with(.data, .fn, .cols = everything(), ...)
#>  
#> row_number(x)
#> 
#> min_rank(x)
#> 
#> dense_rank(x)
#>  
#> rows_insert(
#>   x,
#>   y,
#>   by = NULL,
#>   ...,
#>   conflict = c("error", "ignore"),
#>   copy = FALSE,
#>   in_place = FALSE
#> )
#> 
#> rows_append(x, y, ..., copy = FALSE, in_place = FALSE)
#> 
#> rows_update(
#>   x,
#>   y,
#>   by = NULL,
#>   ...,
#>   unmatched = c("error", "ignore"),
#>   copy = FALSE,
#>   in_place = FALSE
#> )
#> 
#> rows_patch(
#>   x,
#>   y,
#>   by = NULL,
#>   ...,
#>   unmatched = c("error", "ignore"),
#>   copy = FALSE,
#>   in_place = FALSE
#> )
#> 
#> rows_upsert(x, y, by = NULL, ..., copy = FALSE, in_place = FALSE)
#> 
#> rows_delete(
#>   x,
#>   y,
#>   by = NULL,
#>   ...,
#>   unmatched = c("error", "ignore"),
#>   copy = FALSE,
#>   in_place = FALSE
#> )
#>  
#> rowwise(data, ...)
#>  
#> same_src(x, y)
#>  
#> sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)
#> 
#> sample_frac(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...)
#>  NULL 
#> add_count_(x, vars, wt = NULL, sort = FALSE)
#> 
#> add_tally_(x, wt, sort = FALSE)
#> 
#> arrange_(.data, ..., .dots = list())
#> 
#> count_(x, vars, wt = NULL, sort = FALSE, .drop = group_by_drop_default(x))
#> 
#> distinct_(.data, ..., .dots, .keep_all = FALSE)
#> 
#> do_(.data, ..., .dots = list())
#> 
#> filter_(.data, ..., .dots = list())
#> 
#> funs_(dots, args = list(), env = base_env())
#> 
#> group_by_(.data, ..., .dots = list(), add = FALSE)
#> 
#> group_indices_(.data, ..., .dots = list())
#> 
#> mutate_(.data, ..., .dots = list())
#> 
#> tally_(x, wt, sort = FALSE)
#> 
#> transmute_(.data, ..., .dots = list())
#> 
#> rename_(.data, ..., .dots = list())
#> 
#> rename_vars_(vars, args)
#> 
#> select_(.data, ..., .dots = list())
#> 
#> select_vars_(vars, args, include = chr(), exclude = chr())
#> 
#> slice_(.data, ..., .dots = list())
#> 
#> summarise_(.data, ..., .dots = list())
#> 
#> summarize_(.data, ..., .dots = list())
#>  
#> select(.data, ...)
#>  
#> select_all(.tbl, .funs = list(), ...)
#> 
#> rename_all(.tbl, .funs = list(), ...)
#> 
#> select_if(.tbl, .predicate, .funs = list(), ...)
#> 
#> rename_if(.tbl, .predicate, .funs = list(), ...)
#> 
#> select_at(.tbl, .vars, .funs = list(), ...)
#> 
#> rename_at(.tbl, .vars, .funs = list(), ...)
#>  
#> intersect(x, y, ...)
#> 
#> union(x, y, ...)
#> 
#> union_all(x, y, ...)
#> 
#> setdiff(x, y, ...)
#> 
#> setequal(x, y, ...)
#> 
#> symdiff(x, y, ...)
#>  
#> slice(.data, ..., .by = NULL, .preserve = FALSE)
#> 
#> slice_head(.data, ..., n, prop, by = NULL)
#> 
#> slice_tail(.data, ..., n, prop, by = NULL)
#> 
#> slice_min(
#>   .data,
#>   order_by,
#>   ...,
#>   n,
#>   prop,
#>   by = NULL,
#>   with_ties = TRUE,
#>   na_rm = FALSE
#> )
#> 
#> slice_max(
#>   .data,
#>   order_by,
#>   ...,
#>   n,
#>   prop,
#>   by = NULL,
#>   with_ties = TRUE,
#>   na_rm = FALSE
#> )
#> 
#> slice_sample(.data, ..., n, prop, by = NULL, weight_by = NULL, replace = FALSE)
#>  
#> sql(...)
#>  
#> src(subclass, ...)
#> 
#> is.src(x)
#>  
#> src_mysql(
#>   dbname,
#>   host = NULL,
#>   port = 0L,
#>   username = "root",
#>   password = "",
#>   ...
#> )
#> 
#> src_postgres(
#>   dbname = NULL,
#>   host = NULL,
#>   port = NULL,
#>   user = NULL,
#>   password = NULL,
#>   ...
#> )
#> 
#> src_sqlite(path, create = FALSE)
#>  
#> src_local(tbl, pkg = NULL, env = NULL)
#> 
#> src_df(pkg = NULL, env = NULL)
#>  
#> src_tbls(x, ...)
#>  
#> starwars
#>  
#> storms
#>  
#> summarise(.data, ..., .by = NULL, .groups = NULL)
#> 
#> summarize(.data, ..., .by = NULL, .groups = NULL)
#>  
#> summarise_all(.tbl, .funs, ...)
#> 
#> summarise_if(.tbl, .predicate, .funs, ...)
#> 
#> summarise_at(.tbl, .vars, .funs, ..., .cols = NULL)
#> 
#> summarize_all(.tbl, .funs, ...)
#> 
#> summarize_if(.tbl, .predicate, .funs, ...)
#> 
#> summarize_at(.tbl, .vars, .funs, ..., .cols = NULL)
#>  
#> summarise_each(tbl, funs, ...)
#> 
#> summarise_each_(tbl, funs, vars)
#> 
#> mutate_each(tbl, funs, ...)
#> 
#> mutate_each_(tbl, funs, vars)
#> 
#> summarize_each(tbl, funs, ...)
#> 
#> summarize_each_(tbl, funs, vars)
#>  
#> tbl(src, ...)
#> 
#> is.tbl(x)
#>  
#> tbl_df(data)
#> 
#> as.tbl(x, ...)
#>  
#> tbl_ptype(.data)
#>  
#> tbl_vars(x)
#> 
#> tbl_nongroup_vars(x)
#>  NULL 
#> top_n(x, n, wt)
#> 
#> top_frac(x, n, wt)
#>  
#> transmute(.data, ...)
#>  
#> vars(...)
#>  
#> with_groups(.data, .groups, .f, ...)
#>  
#> with_order(order_by, fun, x, ...)
if (FALSE) {
forgot_usg(dplyr, count, write = TRUE)
}